Refactory factory boilerplate
