+++ pkg/stanza/operator/input/generate/config.go
@@ -1,0 +1,37 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package generate // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/generate"
+
+import (
+	"go.opentelemetry.io/collector/component"
+	"go.uber.org/zap"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
+)
+
+// Deprecated [v0.97.0] Use Factory.CreateOperator instead.
+func NewConfig(operatorID string) *Config {
+	return &Config{
+		InputConfig: helper.NewInputConfig(operatorID, operatorType.String()),
+	}
+}
+
+// Config is the configuration of a generate input operator.
+type Config struct {
+	helper.InputConfig `mapstructure:",squash"`
+	Entry              entry.Entry `mapstructure:"entry"`
+	Count              int         `mapstructure:"count"`
+	Static             bool        `mapstructure:"static"`
+}
+
+// Deprecated [v0.97.0] Use Factory.CreateOperator instead.
+func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
+	set := component.TelemetrySettings{}
+	if logger != nil {
+		set.Logger = logger.Desugar()
+	}
+	return NewFactory().CreateOperator(&c, set)
+}
