+++ pkg/stanza/operator/input/generate/input.go
@@ -1,0 +1,87 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package generate // import "github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/generate"
+
+import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
+)
+
+// Input is an operator that generates log entries.
+type Input struct {
+	helper.InputOperator
+	entry  entry.Entry
+	count  int
+	static bool
+	wg     sync.WaitGroup
+	cancel context.CancelFunc
+}
+
+// Start will start generating log entries.
+func (g *Input) Start(_ operator.Persister) error {
+	ctx, cancel := context.WithCancel(context.Background())
+	g.cancel = cancel
+
+	g.wg.Add(1)
+	go func() {
+		defer g.wg.Done()
+		i := 0
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			default:
+			}
+
+			entry := g.entry.Copy()
+			if !g.static {
+				entry.Timestamp = time.Now()
+			}
+			g.Write(ctx, entry)
+
+			i++
+			if i == g.count {
+				return
+			}
+		}
+	}()
+
+	return nil
+}
+
+// Stop will stop generating logs.
+func (g *Input) Stop() error {
+	g.cancel()
+	g.wg.Wait()
+	return nil
+}
+
+func recursiveMapInterfaceToMapString(m any) any {
+	switch m := m.(type) {
+	case map[string]any:
+		newMap := make(map[string]any)
+		for k, v := range m {
+			newMap[k] = recursiveMapInterfaceToMapString(v)
+		}
+		return newMap
+	case map[any]any:
+		newMap := make(map[string]any)
+		for k, v := range m {
+			str, ok := k.(string)
+			if !ok {
+				str = fmt.Sprintf("%v", k)
+			}
+			newMap[str] = recursiveMapInterfaceToMapString(v)
+		}
+		return newMap
+	default:
+		return m
+	}
+}
