+++ pkg/stanza/operator/input/journald/input.go
@@ -39,8 +39,7 @@ func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
 	return NewFactory().CreateOperator(&c, set)
 }
 
-// CreateOperator creates a new operator.
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	inputOperator, err := helper.NewInput(c.InputConfig, set)
 	if err != nil {
