+++ pkg/stanza/operator/parser/jsonarray/factory.go
@@ -28,27 +28,18 @@ func init() {
 	}
 }
 
-type factory struct{}
-
-// NewFactory creates a factory.
+// NewFactory creates a new factory.
 func NewFactory() operator.Factory {
-	return &factory{}
+	return operator.NewFactory(operatorType, newDefaultConfig, createOperator)
 }
 
-// Type gets the type of the operator.
-func (f *factory) Type() component.Type {
-	return operatorType
-}
-
-// NewDefaultConfig creates the default configuration.
-func (f *factory) NewDefaultConfig(operatorID string) component.Config {
+func newDefaultConfig(operatorID string) component.Config {
 	return &Config{
 		ParserConfig: helper.NewParserConfig(operatorID, operatorType.String()),
 	}
 }
 
-// CreateOperator creates a parser.
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	parserOperator, err := helper.NewParser(c.ParserConfig, set)
 	if err != nil {
