+++ pkg/stanza/operator/parser/keyvalue/factory.go
@@ -18,28 +18,19 @@ func init() {
 	operator.RegisterFactory(NewFactory())
 }
 
-type factory struct{}
-
-// NewFactory creates a factory.
+// NewFactory creates a new factory.
 func NewFactory() operator.Factory {
-	return &factory{}
+	return operator.NewFactory(operatorType, newDefaultConfig, createOperator)
 }
 
-// Type gets the type of the operator.
-func (f *factory) Type() component.Type {
-	return operatorType
-}
-
-// CreateOperator creates a parser.
-func (f *factory) NewDefaultConfig(operatorID string) component.Config {
+func newDefaultConfig(operatorID string) component.Config {
 	return &Config{
 		ParserConfig: helper.NewParserConfig(operatorID, operatorType.String()),
 		Delimiter:    "=",
 	}
 }
 
-// CreateOperator creates an operator.
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	parserOperator, err := helper.NewParser(c.ParserConfig, set)
 	if err != nil {
