+++ pkg/stanza/operator/parser/trace/factory.go
@@ -16,28 +16,19 @@ func init() {
 	operator.RegisterFactory(NewFactory())
 }
 
-type factory struct{}
-
-// NewFactory creates a factory.
+// NewFactory creates a new factory.
 func NewFactory() operator.Factory {
-	return &factory{}
+	return operator.NewFactory(operatorType, newDefaultConfig, createOperator)
 }
 
-// Type gets the type of the operator.
-func (f *factory) Type() component.Type {
-	return operatorType
-}
-
-// NewDefaultConfig creates the default configuration.
-func (f *factory) NewDefaultConfig(operatorID string) component.Config {
+func newDefaultConfig(operatorID string) component.Config {
 	return &Config{
 		TransformerConfig: helper.NewTransformerConfig(operatorID, operatorType.String()),
 		TraceParser:       helper.NewTraceParser(),
 	}
 }
 
-// CreateOperator creates a trace parser.
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	transformerOperator, err := helper.NewTransformer(c.TransformerConfig, set)
 	if err != nil {
