+++ pkg/stanza/operator/transformer/recombine/factory.go
@@ -23,20 +23,12 @@ func init() {
 	operator.RegisterFactory(NewFactory())
 }
 
-type factory struct{}
-
-// NewFactory creates a factory
+// NewFactory creates a new factory.
 func NewFactory() operator.Factory {
-	return &factory{}
+	return operator.NewFactory(operatorType, newDefaultConfig, createOperator)
 }
 
-// Type gets the type of the operator
-func (f *factory) Type() component.Type {
-	return operatorType
-}
-
-// NewDefaultConfig creates the default configuration
-func (f *factory) NewDefaultConfig(operatorID string) component.Config {
+func newDefaultConfig(operatorID string) component.Config {
 	return &Config{
 		TransformerConfig: helper.NewTransformerConfig(operatorID, operatorType.String()),
 		MaxBatchSize:      1000,
@@ -48,8 +40,7 @@ func (f *factory) NewDefaultConfig(operatorID string) component.Config {
 	}
 }
 
-// CreateOperator creates an operator
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	transformer, err := helper.NewTransformer(c.TransformerConfig, set)
 	if err != nil {
