+++ pkg/stanza/operator/transformer/remove/factory.go
@@ -19,27 +19,18 @@ func init() {
 	operator.RegisterFactory(NewFactory())
 }
 
-type factory struct{}
-
-// NewFactory creates a factory
+// NewFactory creates a new factory.
 func NewFactory() operator.Factory {
-	return &factory{}
+	return operator.NewFactory(operatorType, newDefaultConfig, createOperator)
 }
 
-// Type gets the type of the operator
-func (f *factory) Type() component.Type {
-	return operatorType
-}
-
-// NewDefaultConfig creates the default configuration
-func (f *factory) NewDefaultConfig(operatorID string) component.Config {
+func newDefaultConfig(operatorID string) component.Config {
 	return &Config{
 		TransformerConfig: helper.NewTransformerConfig(operatorID, operatorType.String()),
 	}
 }
 
-// CreateOperator creates an operator
-func (f *factory) CreateOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
+func createOperator(cfg component.Config, set component.TelemetrySettings) (operator.Operator, error) {
 	c := cfg.(*Config)
 	transformerOperator, err := helper.NewTransformer(c.TransformerConfig, set)
 	if err != nil {
